refactor: clean up user profile layout
