+++ apps/web/components/buttons/module-state-guide.tsx
@@ -1,6 +1,6 @@
 import { useAppDispatch, useAppSelector, r } from '@/store/redux'
 import { CloseButton } from '@/ui/buttons'
-import { flatten } from '@/utils/tailwind'
+import { cc } from '@/utils/tailwind'
 
 export function ModuleStateGuide() {
   const show = useAppSelector((s) => s.modal.showModuleStateGuide)
@@ -8,9 +8,7 @@ export function ModuleStateGuide() {
   return show ? (
     <div className="absolute right-10 bottom-10 select-none">
       <div className="flex flex-row">
-        <div
-          className={flatten('px-4 py-3', 'bg-white rounded-lg', 'shadow-xl')}
-        >
+        <div className={cc('px-4 py-3', 'bg-white rounded-lg', 'shadow-xl')}>
           <h5 className="flex flex-row h-6 items-center">
             <span className="flex-1 mb-0">Module state guide</span>
             <div>
