--- apps/web/components/user-profile/panels/index.tsx
@@ -1,2 +1,0 @@
-export * from './account'
-export * from './public-profile'
