+++ apps/web/types/index.d.ts
@@ -78,13 +78,7 @@ export type Pages = {
   Degrees: 'main' | 'add-new' | 'edit'
   Graphs: 'main' | 'add-new' | 'edit'
   Modules: 'main' | 'add-doing' | 'add-done'
-  UserProfile:
-    | 'public-profile'
-    | 'account'
-    | 'graphs'
-    | 'modules'
-    | 'degrees'
-    | 'debug'
+  UserProfile: 'graphs' | 'modules' | 'degrees'
 }
 
 export type ContextMenuType =
