+++ apps/web/utils/tailwind.ts
@@ -1,3 +1,4 @@
-export function flatten(...args: (string | undefined)[]) {
+/** combine css */
+export function cc(...args: (string | undefined)[]) {
   return args.filter(Boolean).join(' ')
 }
